+++ src/components/StarRating/StarRating.js
@@ -5,16 +5,29 @@ import Icon from 'react-native-vector-icons/MaterialIcons';
 
 import utils from '../../utils';
 import styles from './styles';
+import Touchable from '../Touchable';
 
 const propTypes = {
   rating: PropTypes.number,
+  handlePress: PropTypes.func,
+  androidRipple: PropTypes.bool,
+  androidRippleColor: PropTypes.string,
+  androidRippleBorderless: PropTypes.bool,
   iconStyle: Text.propTypes.style,
   style: ViewPropTypes.style,
 };
 
 const defaultProps = {};
 
-const StarRating = ({ rating, iconStyle, style }) => {
+const StarRating = ({
+  rating,
+  handlePress,
+  androidRipple,
+  androidRippleColor,
+  androidRippleBorderless,
+  iconStyle,
+  style,
+}) => {
   const maxStars = 5;
   const starsArray = [];
   for (let i = 0; i < maxStars; i += 1) {
@@ -33,8 +46,19 @@ const StarRating = ({ rating, iconStyle, style }) => {
 
   return (
     <View style={[styles.container, style]}>
-      {starsArray.map((item) => {
-        return <Icon key={item.id} name={item.iconName} style={[styles.icon, iconStyle]} />;
+      {starsArray.map((item, index) => {
+        return (
+          <Touchable
+            key={item.id}
+            onPress={() => handlePress(index + 1)} // ratings of 1 - 5
+            androidRipple={androidRipple}
+            androidRippleColor={androidRippleColor}
+            androidRippleBorderless={androidRippleBorderless}
+            disabled={!handlePress}
+          >
+            <Icon name={item.iconName} style={[styles.icon, iconStyle]} />
+          </Touchable>
+        );
       })}
     </View>
   );