+++ src/components/StarRatingInput/index.js
@@ -1,3 +1,0 @@
-import StarRatingInput from './StarRatingInput';
-
-export default StarRatingInput;